feat(model)!: Rework Notifier job configuration
The job configuration now only contains the properties that can be set by
end users. Other configuration options are specified in the admin
configuration.

Adapt the UI accordingly.
