--- api/v1/model/src/commonMain/kotlin/MailNotificationConfiguration.kt
@@ -1,38 +1,0 @@
-/*
- * Copyright (C) 2024 The ORT Server Authors (See <https://github.com/eclipse-apoapsis/ort-server/blob/main/NOTICE>)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * License-Filename: LICENSE
- */
-
-package org.eclipse.apoapsis.ortserver.api.v1.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class MailNotificationConfiguration(
-    val recipientAddresses: List<String>? = null,
-    val mailServerConfiguration: MailServerConfiguration? = null
-)
-
-@Serializable
-data class MailServerConfiguration(
-    val hostName: String,
-    val port: Int,
-    val username: String,
-    val password: String,
-    val useSsl: Boolean,
-    val fromAddress: String
-)
